+++ langchain/src/chat_models/tests/chatopenai.int.test.ts
@@ -611,3 +611,156 @@ test("ChatOpenAI should not reuse cache if function call args have changed", asy
   lookupSpy.mockRestore();
   updateSpy.mockRestore();
 });
+
+test("Test ChatOpenAI token usage reporting for streaming function calls", async () => {
+  let streamingTokenUsed = -1;
+  let nonStreamingTokenUsed = -1;
+
+  const humanMessage = "What a beautiful day!";
+  const extractionFunctionSchema = {
+    name: "extractor",
+    description: "Extracts fields from the input.",
+    parameters: {
+      type: "object",
+      properties: {
+        tone: {
+          type: "string",
+          enum: ["positive", "negative"],
+          description: "The overall tone of the input",
+        },
+        word_count: {
+          type: "number",
+          description: "The number of words in the input",
+        },
+        chat_response: {
+          type: "string",
+          description: "A response to the human's input",
+        },
+      },
+      required: ["tone", "word_count", "chat_response"],
+    },
+  };
+
+  const streamingModel = new ChatOpenAI({
+    modelName: "gpt-3.5-turbo",
+    streaming: true,
+    maxRetries: 10,
+    maxConcurrency: 10,
+    temperature: 0,
+    topP: 0,
+    callbacks: [
+      {
+        handleLLMEnd: async (output) => {
+          streamingTokenUsed = output.llmOutput?.tokenUsage?.totalTokens;
+        },
+        handleLLMError: async (err) => {
+          console.error(err);
+        },
+      },
+    ],
+  }).bind({
+    functions: [extractionFunctionSchema],
+    function_call: { name: "extractor" },
+  });
+
+  const nonStreamingModel = new ChatOpenAI({
+    modelName: "gpt-3.5-turbo",
+    streaming: false,
+    maxRetries: 10,
+    maxConcurrency: 10,
+    temperature: 0,
+    topP: 0,
+    callbacks: [
+      {
+        handleLLMEnd: async (output) => {
+          nonStreamingTokenUsed = output.llmOutput?.tokenUsage?.totalTokens;
+        },
+        handleLLMError: async (err) => {
+          console.error(err);
+        },
+      },
+    ],
+  }).bind({
+    functions: [extractionFunctionSchema],
+    function_call: { name: "extractor" },
+  });
+
+  const [nonStreamingResult, streamingResult] = await Promise.all([
+    nonStreamingModel.invoke([new HumanMessage(humanMessage)]),
+    streamingModel.invoke([new HumanMessage(humanMessage)]),
+  ]);
+
+  if (
+    nonStreamingResult.additional_kwargs.function_call?.arguments &&
+    streamingResult.additional_kwargs.function_call?.arguments
+  ) {
+    const nonStreamingArguments = JSON.stringify(
+      JSON.parse(nonStreamingResult.additional_kwargs.function_call.arguments)
+    );
+    const streamingArguments = JSON.stringify(
+      JSON.parse(streamingResult.additional_kwargs.function_call.arguments)
+    );
+    if (nonStreamingArguments === streamingArguments) {
+      expect(streamingTokenUsed).toEqual(nonStreamingTokenUsed);
+    }
+  }
+
+  expect(streamingTokenUsed).toBeGreaterThan(-1);
+});
+
+test("Test ChatOpenAI token usage reporting for streaming calls", async () => {
+  let streamingTokenUsed = -1;
+  let nonStreamingTokenUsed = -1;
+  const question = "What is the color of the night sky?";
+
+  const streamingModel = new ChatOpenAI({
+    modelName: "gpt-3.5-turbo",
+    streaming: true,
+    maxRetries: 10,
+    maxConcurrency: 10,
+    temperature: 0,
+    topP: 0,
+    callbacks: [
+      {
+        handleLLMEnd: async (output) => {
+          streamingTokenUsed = output.llmOutput?.tokenUsage?.totalTokens;
+        },
+        handleLLMError: async (err) => {
+          console.error(err);
+        },
+      },
+    ],
+  });
+
+  const nonStreamingModel = new ChatOpenAI({
+    modelName: "gpt-3.5-turbo",
+    streaming: false,
+    maxRetries: 10,
+    maxConcurrency: 10,
+    temperature: 0,
+    topP: 0,
+    callbacks: [
+      {
+        handleLLMEnd: async (output) => {
+          nonStreamingTokenUsed = output.llmOutput?.tokenUsage?.totalTokens;
+        },
+        handleLLMError: async (err) => {
+          console.error(err);
+        },
+      },
+    ],
+  });
+
+  const [nonStreamingResult, streamingResult] = await Promise.all([
+    nonStreamingModel.generate([[new HumanMessage(question)]]),
+    streamingModel.generate([[new HumanMessage(question)]]),
+  ]);
+
+  expect(streamingTokenUsed).toBeGreaterThan(-1);
+  if (
+    nonStreamingResult.generations[0][0].text ===
+    streamingResult.generations[0][0].text
+  ) {
+    expect(streamingTokenUsed).toEqual(nonStreamingTokenUsed);
+  }
+});
